backend/requirements.txt
asgiref==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==5.2.13
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==10.4.0
PyJWT==2.12.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3